ml/requirements.txt
lightgbm>=4.0
xgboost>=2.0
catboost>=1.2
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
shap>=0.42
optuna>=3.3
matplotlib>=3.7
joblib>=1.3
python-dotenv>=1.0
